create utils folder to organize code
